feat: Implement resource and region customization via overrides file (#60)
Adds a new override system that enables users to modify & add resource or location slugs, without the need to update the provider. A YAML file can now be added to your project root with some simple configuration to override or add regions or resources.

This enables flexibility in environment where existing naming conventions may disagree with the conventions coded in the provider.

Fixes #59